bank account keeping

+++ src/main/java/com/ilyapolyanskiy/practice/module02/task022/TKRunner.java
@@ -5,8 +5,8 @@ package com.ilyapolyanskiy.practice.module02.task022;
  */
 public class TKRunner {
     public static void main(String[] args) {
-        int balance = 100;
-        int withdrawal = 10;
+        double balance = 100.0;
+        double withdrawal = 10.0;
         TactionKeeper keeper = new TactionKeeper();
         if (keeper.balanceCalc(balance, withdrawal) > 0)
             System.out.println("OK " + keeper.commissionCalc(withdrawal) + " " + keeper.balanceCalc(balance, withdrawal));
